WIP
Test the use of the BulkExecutor library to improve the performance of the
upserts of each documents being mutated as part of a DocumentMigration.

The BulkExecutor library does give a good improvement (~10x over the initial
implementation).

However due to the way dependencies are handled within cake and it's non-support
of binding redirects we have abandoned this work.